% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_f}
\alias{anova_f}
\title{ANOVA f-stat from msg and msr}
\usage{
anova_f(msg, msr)
}
\arguments{
\item{msg}{Mean Squares Group}

\item{msr}{Mean Squares Residual}
}
\value{
f-stat
}
\description{
ANOVA f-stat from msg and msr
}
