% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_msg}
\alias{anova_msg}
\title{ANOVA Mean Squares Group}
\usage{
anova_msg(ssg, k)
}
\arguments{
\item{ssg}{The Sum of Squares Group}

\item{k}{The number of groups}
}
\value{
the mean squares for the group
}
\description{
ANOVA Mean Squares Group
}
