% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_msr}
\alias{anova_msr}
\title{ANOVA Mean Squares Residuals}
\usage{
anova_msr(ssr, n, k)
}
\arguments{
\item{ssr}{the Sum of Squares Residuals}

\item{n}{the sample size}

\item{k}{The number of groups}
}
\value{
the mean squares for the residuals
}
\description{
ANOVA Mean Squares Residuals
}
