% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_mst}
\alias{anova_mst}
\title{ANOVA Mean Squares Total}
\usage{
anova_mst(sst, n)
}
\arguments{
\item{sst}{the Sum of Squares Total}

\item{n}{the sample size}
}
\value{
the mean squares Total
}
\description{
ANOVA Mean Squares Total
}
