% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-distributions.R
\name{binom_dist}
\alias{binom_dist}
\title{Binomial Distribution}
\usage{
binom_dist(n, prob)
}
\arguments{
\item{n}{number of outcomes}

\item{prob}{the probability}
}
\value{
E(X), Var(X), sd(X)
}
\description{
Calculates the Expected mean, variance and standard deviation for
a binomial distribution
}
\examples{
binom_dist(5, 0.584)
}
