% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-distributions.R
\name{central_limit}
\alias{central_limit}
\title{Central Limit Theorem to Standardize a normal distribution}
\usage{
central_limit(x, mu, sigma, n)
}
\arguments{
\item{x}{the sample mean.}

\item{mu}{the population mean}

\item{sigma}{the population standard deviation}

\item{n}{the sample size}
}
\value{
standardized z-stat
}
\description{
If the population is not normally distributed, for sufficiently large n,
we use central limit theorem.
}
