% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-square.R
\name{chisq_gof}
\alias{chisq_gof}
\title{Chi-Square Goodness of Fit Test}
\usage{
chisq_gof(fo, p, k = 0)
}
\arguments{
\item{fo}{the observed frequencies.}

\item{p}{the probability of each outcome.}

\item{k}{the number of parameters estimated using the data.}
}
\value{
the list of stats.
}
\description{
Calculates the frequency table, and the statistics, prints in a nice format,
and returns a list for easy access of the stats.
}
\examples{
chisq_gof(c(185, 190, 210, 205, 195, 215), 1 / 6)
chisq_gof(c(Low = 100, Med = 255, High = 45), c(0.2, 0.7, 0.1))
}
