% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-square.R
\name{chisq_indep_df}
\alias{chisq_indep_df}
\title{Chi-Square Independence Degrees of Freedom}
\usage{
chisq_indep_df(num.rows, num.cols)
}
\arguments{
\item{num.rows}{Number of rows}

\item{num.cols}{Number of columns}
}
\value{
degrees of freedom
}
\description{
Chi-Square Independence Degrees of Freedom
}
