% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-square.R
\name{chisq_indep_expected}
\alias{chisq_indep_expected}
\title{Chi-Square Independence Expected Frequency}
\usage{
chisq_indep_expected(RowTotal, ColTotal, GrandTotal)
}
\arguments{
\item{RowTotal}{the row total}

\item{ColTotal}{the column total}

\item{GrandTotal}{the grand total}
}
\value{
the expected frequency for the specified cell
}
\description{
Chi-Square Independence Expected Frequency
}
