% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation_df}
\alias{correlation_df}
\title{Correlation Between Two Variables DF}
\usage{
correlation_df(n)
}
\arguments{
\item{n}{sample size}
}
\value{
Degrees of Freedom
}
\description{
Correlation Between Two Variables DF
}
\examples{
correlation_df(20)
}
