% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation_se}
\alias{correlation_se}
\title{Correlation Standard Error}
\usage{
correlation_se(r, n)
}
\arguments{
\item{r}{the sample correlation coefficient can be found with \code{cor(x, y)}}

\item{n}{the sample size}
}
\value{
The correlation standard error
}
\description{
Correlation Standard Error
}
\seealso{
\code{\link[stats:cor]{stats::cor()}}

\code{\link[stats:cor.test]{stats::cor.test()}}
}
