% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation_t}
\alias{correlation_t}
\title{Correlation t-test}
\usage{
correlation_t(r, n, tail = 2, p = 0, conf = 0.95)
}
\arguments{
\item{r}{the sample correlation coefficient can be found with \code{cor(x, y)}}

\item{n}{the sample size}

\item{tail}{specify either \code{1} or \code{2} tail}

\item{p}{the population correlation coefficient}

\item{conf}{the confidence}
}
\value{
the list of stats.
}
\description{
Uses Fisher's Z Transformation for Confidence Interval
}
\examples{
correlation_t(0.6642512, 20, 2)
}
\seealso{
\code{\link[stats:cor]{stats::cor()}}

\code{\link[stats:cor.test]{stats::cor.test()}}
}
