% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-distributions.R
\name{discrete_dist}
\alias{discrete_dist}
\title{Discrete Probability Distribution}
\usage{
discrete_dist(x, prob)
}
\arguments{
\item{x}{the outcome}

\item{prob}{the probability of the outcome}
}
\value{
the list of stats
}
\description{
Calculates the Expected mean, variance and standard deviation for
a discrete probability distribution
}
\examples{
discrete_dist(0:3, c(0.21, 0.45, 0.23, 0.11))
}
