% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{f_test}
\alias{f_test}
\title{ANOVA f-test}
\usage{
f_test(f, df1, df2)
}
\arguments{
\item{f}{the f-stat}

\item{df1}{group degrees of freedom (k - 1)}

\item{df2}{residuals degrees of freedom (n - k)}
}
\value{
p-value
}
\description{
ANOVA f-test
}
\seealso{
\code{\link[stats:Fdist]{stats::pf()}}

\code{\link[stats:aov]{stats::aov()}}

\code{aov(Y ~ X, data) # One-Way}

\code{aov(Y ~ X1 * X2, data) # Two-Way}

\code{\link[stats:pairwise.t.test]{stats::pairwise.t.test()}}

\code{pairwise.t.test(x, g, p.adjust.method = "none")}

\code{pairwise.t.test(x, g, p.adjust.method = "bonferroni") # significance = 0.05/numPairs}

\code{\link[stats:TukeyHSD]{stats::TukeyHSD()}}

\code{TukeyHSD(fit)}
}
