% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-square.R
\name{generate_frame}
\alias{generate_frame}
\title{Generate data.frame}
\usage{
generate_frame(...)
}
\arguments{
\item{...}{The observed frequencies (fo) in the form of \cr
Col1 = c(Row1 = 52, Row2 = 61), Col2 = c(Row1 = 22, Row2 = 39)}
}
\value{
the new data.frame
}
\description{
Generates a data.frame from observed frequencies.
}
\examples{
generate_frame(Nicotine = c(No = 148, Yes = 52), Placebo = c(No = 182, Yes = 18))
generate_frame(Home = c("2018" = 391, "2020" = 454), Shared = c("2018" = 185, "2020" = 139))
}
