% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{interval}
\alias{interval}
\title{Confidence Interval}
\usage{
interval(x, moe)
}
\arguments{
\item{x}{the value being intervaled}

\item{moe}{the margin of error}
}
\value{
\code{x +- moe}
}
\description{
Calculates the confidence interval of a value.
}
