% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{logistic_groups}
\alias{logistic_groups}
\title{Logistic Regression to Compare Odds Between Groups}
\usage{
logistic_groups(b0, b1)
}
\arguments{
\item{b0}{the estimate intercept}

\item{b1}{the estimate of the variable}
}
\value{
Odds Ratio between groups
}
\description{
Logistic Regression to Compare Odds Between Groups
}
\examples{
logistic_groups(-2.3136, 1.2677)
}
