% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{logistic_or}
\alias{logistic_or}
\title{Odds Ratio (OR)}
\usage{
logistic_or(p1, p2)
}
\arguments{
\item{p1}{Probability of event 1}

\item{p2}{Probability of event 2}
}
\value{
Odds Ratio
}
\description{
Odds Ratio (OR)
}
