% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{logistic_phat}
\alias{logistic_phat}
\title{Logistic Regression phat}
\usage{
logistic_phat(b0, b1, x)
}
\arguments{
\item{b0}{the estimate intercept}

\item{b1}{the estimate of the variable}

\item{x}{the value being predicted}
}
\value{
estimated probability.
}
\description{
Logistic Regression phat
}
\examples{
logistic_phat(-12.5974, 0.7630, 15)
}
