% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-regression.R
\name{lr_df}
\alias{lr_df}
\title{Linear Regression Degrees of Freedom}
\usage{
lr_df(n, k)
}
\arguments{
\item{n}{the sample size}

\item{k}{the number of groups}
}
\value{
the df
}
\description{
Linear Regression Degrees of Freedom
}
