% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-regression.R
\name{lr_moe}
\alias{lr_moe}
\title{Linear Regression Margin of Error}
\usage{
lr_moe(df, se, conf = 0.95)
}
\arguments{
\item{df}{Degrees of Freedom}

\item{se}{Standard Error}

\item{conf}{Confidence}
}
\value{
Margin of Error
}
\description{
Linear Regression Margin of Error
}
