% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-regression.R
\name{mlr_adjusted_R2}
\alias{mlr_adjusted_R2}
\title{Adjusted R-Squared Value}
\usage{
mlr_adjusted_R2(R2, n, k)
}
\arguments{
\item{R2}{the non-adjusted R-Squared}

\item{n}{the sample size}

\item{k}{the number of independent variables}
}
\value{
adjusted R-Squared
}
\description{
Calculates the adjusted R-Squared value for multiple Linear Regression
}
\examples{
mlr_adjusted_R2(0.78, 20, 2)
}
