% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{mode_stats}
\alias{mode_stats}
\title{Mode of from values}
\usage{
mode_stats(x)
}
\arguments{
\item{x}{values}
}
\value{
the mode of the values
}
\description{
Mode of from values
}
\examples{
mode_stats(c(80, 70, 66, 50, 66, 74, 78, 58))
}
