% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-templates.R
\name{moe_base}
\alias{moe_base}
\title{moe base}
\arguments{
\item{conf}{The confidence}

\item{df}{The degrees of freedom}

\item{se}{The standard error}
}
\value{
The margin of Error
}
\description{
Calculates the margin of error
}
