% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ni}
\alias{ni}
\alias{\%ni\%}
\title{Not Value Matching}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\description{
Negation of \code{\%in\%}
}
\examples{
1 \%ni\% c(2, 3) # TRUE
"B" \%ni\% c("A", "B", "C") # FALSE
}
