% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{one_prop_se}
\alias{one_prop_se}
\title{One Sample Proportion Standard Error}
\usage{
one_prop_se(phat, n)
}
\arguments{
\item{phat}{Sample proportion}

\item{n}{Sample size}
}
\value{
Standard Error
}
\description{
One Sample Proportion Standard Error
}
\examples{
one_prop_se(10 / 15, 15)
}
