% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{one_prop_z}
\alias{one_prop_z}
\title{One Sample Proportion z-test}
\usage{
one_prop_z(phat, n, p = 0, tail = 2, conf = 0.95)
}
\arguments{
\item{phat}{Sample proportion}

\item{n}{Sample size}

\item{p}{population proportion}

\item{tail}{specify either \code{1} or \code{2} tail}

\item{conf}{the confidence}
}
\value{
the list of stats.
}
\description{
Calculates the statistics, prints in a nice format,
and returns a list for easy access of the stats.
}
