% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-sample.R
\name{one_sample_assumptions}
\alias{one_sample_assumptions}
\title{One Sample t-test Assumptions}
\usage{
one_sample_assumptions()
}
\description{
Prints out the assumptions.
}
\examples{
one_sample_assumptions()
}
