% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-sample.R
\name{one_sample_se}
\alias{one_sample_se}
\title{One Sample Standard Error}
\usage{
one_sample_se(s, n)
}
\arguments{
\item{s}{the sample standard deviation}

\item{n}{the sample size}
}
\value{
students se
}
\description{
Calculates the students standard error
}
