% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-sample.R
\name{one_sample_t}
\alias{one_sample_t}
\title{One Sample Student's t-test}
\usage{
one_sample_t(x, s, n, mu = 0, tail = 2, conf = 0.95)
}
\arguments{
\item{x}{the sample mean}

\item{s}{the sample standard deviation}

\item{n}{the sample size}

\item{mu}{the population mean}

\item{tail}{specify either \code{1} or \code{2} tail}

\item{conf}{the confidence}
}
\value{
the list of stats.
}
\description{
Calculates the statistics, prints in a nice format,
and returns a list for easy access of the stats.
}
\seealso{
\code{\link[stats:t.test]{stats::t.test()}}

\code{t.test(data$X)}
}
