% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{outliers}
\alias{outliers}
\title{Outliers}
\usage{
outliers(Q1, Q3, obs = NULL)
}
\arguments{
\item{Q1}{the first quartile}

\item{Q3}{the third quartile}

\item{obs}{the observation}
}
\description{
Displays the left and right bounds for outliers.
If an observation is provided displays whether that observation is an outlier.
}
\examples{
outliers(6.45, 10.23)
outliers(6.45, 10.23, obs = 0.5)
outliers(6.45, 10.23, obs = 16)
outliers(6.45, 10.23, obs = 12)
}
