% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample.R
\name{pooled_S2p}
\alias{pooled_S2p}
\title{Pooled Variance}
\usage{
pooled_S2p(s1, n1, s2, n2)
}
\arguments{
\item{s1}{1st sample SD}

\item{n1}{1st sample size}

\item{s2}{2nd sample SD}

\item{n2}{2nd sample size}
}
\value{
Pooled variance
}
\description{
Pooled Variance
}
