% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample.R
\name{pooled_se}
\alias{pooled_se}
\title{Pooled Standard Error}
\usage{
pooled_se(s1, n1, s2, n2)
}
\arguments{
\item{s1}{1st sample SD}

\item{n1}{1st sample size}

\item{s2}{2nd sample SD}

\item{n2}{2nd sample size}
}
\value{
Pooled se(x1 - x2)
}
\description{
Pooled Standard Error
}
