% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample.R
\name{pooled_t}
\alias{pooled_t}
\title{Pooled t-test SD's Equal}
\usage{
pooled_t(x1, s1, n1, x2, s2, n2, tail, conf = 0.95)
}
\arguments{
\item{x1}{1st sample mean}

\item{s1}{1st sample SD}

\item{n1}{1st sample size}

\item{x2}{2nd sample mean}

\item{s2}{2nd sample SD}

\item{n2}{2nd sample size}

\item{tail}{specify either \code{1} or \code{2} tail}

\item{conf}{the confidence}
}
\value{
the list of stats.
}
\description{
Calculates the statistics, prints in a nice format,
and returns a list for easy access of the stats.
}
\seealso{
\code{stats::t.test(Y ~ X, data, var.equal = TRUE)}
}
