% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{population_sd}
\alias{population_sd}
\title{Population Standard Deviation}
\usage{
population_sd(xi, mu, N)
}
\arguments{
\item{xi}{population}

\item{mu}{population mean}

\item{N}{population size}
}
\value{
Population Standard Deviation
}
\description{
Population Standard Deviation
}
