% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{prop_moe}
\alias{prop_moe}
\title{Proportion Margin of Error}
\usage{
prop_moe(se, conf = 0.95)
}
\arguments{
\item{se}{The standard error}

\item{conf}{The confidence}
}
\value{
The margin of Error
}
\description{
Calculates the margin of error
}
