% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{rank_assumptions}
\alias{rank_assumptions}
\title{Signed-Rank Assumptions (Maybe Sum-Rank too?)}
\usage{
rank_assumptions()
}
\description{
Prints out the assumptions.
}
\examples{
rank_assumptions()
}
