% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{rank_sum_expected}
\alias{rank_sum_expected}
\title{Rank-Sum Expected Mean With Normal Approximation}
\usage{
rank_sum_expected(n1, n2)
}
\arguments{
\item{n1}{the sample size whose ranks are summing}

\item{n2}{the sample size of the other group}
}
\value{
the expected mean
}
\description{
Rank-Sum Expected Mean With Normal Approximation
}
