% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{rank_sum_sd}
\alias{rank_sum_sd}
\title{Rank-Sum standard Deviation With Normal Approximation}
\usage{
rank_sum_sd(n1, n2)
}
\arguments{
\item{n1}{the sample size whose ranks are summing}

\item{n2}{the sample size of the other group}
}
\value{
the standard deviation
}
\description{
Rank-Sum standard Deviation With Normal Approximation
}
