% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{sample_sd}
\alias{sample_sd}
\title{Sample Standard Deviation}
\usage{
sample_sd(xi)
}
\arguments{
\item{xi}{sample}
}
\value{
Sample Standard Deviation
}
\description{
Sample Standard Deviation
}
\seealso{
\code{\link[stats:sd]{stats::sd()}}
}
