% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-distributions.R
\name{sampling_dist_mean}
\alias{sampling_dist_mean}
\title{Sampling Distribution of the Sample Mean}
\usage{
sampling_dist_mean(mu, sigma, n)
}
\arguments{
\item{mu}{population mean}

\item{sigma}{population sd}

\item{n}{random sample size}
}
\value{
E(Xbar), Var(Xbar), sd(Xbar)
}
\description{
Sampling Distribution of the Sample Mean
}
\examples{
sampling_dist_mean(50, 8, 4)
}
