% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-distributions.R
\name{sampling_dist_prop}
\alias{sampling_dist_prop}
\title{Sampling Distribution of the Sample Proportions}
\usage{
sampling_dist_prop(p, n)
}
\arguments{
\item{p}{population proportion}

\item{n}{random sample size}
}
\value{
E(phat), Var(phat), sd(phat)
}
\description{
Sampling Distribution of the Sample Proportions
}
\examples{
sampling_dist_prop(0.1, 10)
}
