% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-distributions.R
\name{sampling_dist_prop_norm}
\alias{sampling_dist_prop_norm}
\title{Transform phat ~ Normal(p, sd(phat)) into standard normal form.}
\usage{
sampling_dist_prop_norm(phat, p, n)
}
\arguments{
\item{phat}{the sample proportion}

\item{p}{population proportion}

\item{n}{random sample size}
}
\value{
the standardized z-stat.
}
\description{
Transform phat ~ Normal(p, sd(phat)) into standard normal form.
}
\examples{
sampling_dist_prop_norm(0.3, 0.1, 10)
}
