% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{signed_rank}
\alias{signed_rank}
\title{Signed-Rank test with Normal Approximation}
\usage{
signed_rank(S, n, tail)
}
\arguments{
\item{S}{sum of ranks corresponding to either positive or negative differences}

\item{n}{the sample size}

\item{tail}{specify either \code{1} or \code{2} tail}
}
\value{
the list of stats.
}
\description{
Calculates the statistics, prints in a nice format,
and returns a list for easy access of the stats.
}
\examples{
signed_rank(96, 15, 1)
}
