% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{signed_rank_expected}
\alias{signed_rank_expected}
\title{Signed-Rank Expected Mean with Normal Approximation}
\usage{
signed_rank_expected(n)
}
\arguments{
\item{n}{the sample size}
}
\value{
expected mean
}
\description{
Signed-Rank Expected Mean with Normal Approximation
}
