% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{signed_rank_sd}
\alias{signed_rank_sd}
\title{Signed-Rank Standard Deviation with Normal Approximation}
\usage{
signed_rank_sd(n)
}
\arguments{
\item{n}{the sample size}
}
\value{
standard deviation
}
\description{
Signed-Rank Standard Deviation with Normal Approximation
}
