% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-regression.R
\name{slr_se}
\alias{slr_se}
\title{Simple Linear Regression Standard Error}
\usage{
slr_se(b, t, B = 0)
}
\arguments{
\item{b}{the sample estimate}

\item{t}{the t-stat}

\item{B}{the population estimate}
}
\value{
the se
}
\description{
Simple Linear Regression Standard Error
}
