% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-regression.R
\name{slr_t}
\alias{slr_t}
\title{Simple Linear Regression t-stat}
\usage{
slr_t(b, se, B = 0)
}
\arguments{
\item{b}{the sample estimate}

\item{se}{the standard error}

\item{B}{the population estimate}
}
\value{
the t-stat
}
\description{
Simple Linear Regression t-stat
}
\seealso{
\code{\link[stats:lm]{stats::lm()}}

\code{lm(Y ~ X, data)}
}
