% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-distributions.R
\name{stand_normal}
\alias{stand_normal}
\title{Standardize a normal distribution}
\usage{
stand_normal(x, mu, sigma)
}
\arguments{
\item{x}{value}

\item{mu}{population mean}

\item{sigma}{population standard deviation}
}
\value{
standardized z-stat
}
\description{
Transform a normal distribution to a standard normal distribution
}
