% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat1201.R
\docType{package}
\name{stat1201}
\alias{stat1201}
\title{STAT1201 Equation Library}
\description{
All the equations you need for the
\href{https://my.uq.edu.au/programs-courses/course.html?course_code=stat1201}{STAT1201} Final
}
\seealso{
\href{https://jedwillick.github.io/stat1201/}{Documentation}

\href{https://github.com/jedwillick/stat1201}{Source}
}
