% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{stats_print}
\alias{stats_print}
\title{Print Stats}
\usage{
stats_print(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[base:list]{base::list}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
the list of stats.
}
\description{
Prints the statistics in a nice format, and returns the values as list.
}
