% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-sample.R
\name{students_se}
\alias{students_se}
\title{Student's Standard Error}
\usage{
students_se(s, n)
}
\arguments{
\item{s}{the sample standard deviation}

\item{n}{the sample size}
}
\value{
students se
}
\description{
Calculates the students standard error
}
