% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-sample.R
\name{students_t}
\alias{students_t}
\title{Student's One Sample t-test}
\usage{
students_t(mu, x, s, n, tail, conf = 0.95)
}
\arguments{
\item{mu}{the population mean}

\item{x}{the sample mean}

\item{s}{the sample standard deviation}

\item{n}{the sample size}

\item{tail}{specify either \code{1} or \code{2} tail}

\item{conf}{the confidence}
}
\value{
the list of stats.
}
\description{
Calculates the statistics, prints in a nice format,
and returns a list for easy access of the stats.
}
\seealso{
\code{stats::t.test(data$X)}
}
