% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{t_crit}
\alias{t_crit}
\title{t Critical}
\usage{
t_crit(conf, df)
}
\arguments{
\item{conf}{The confidence as a decimal.}

\item{df}{the degrees of freedom}
}
\value{
the t Critical value
}
\description{
Calculates the t Critical value from a confidence.
}
\examples{
t_crit(0.95, 2) # 95\% Confident
t_crit(0.90, 2) # 90\% Confident
}
