% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{t_test}
\alias{t_test}
\title{t-test}
\usage{
t_test(t, df, tail)
}
\arguments{
\item{t}{the t stat}

\item{df}{the degrees of freedom}

\item{tail}{specify either \code{1} or \code{2} tail}
}
\value{
p value
}
\description{
Performs a t-test.
}
