% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-templates.R
\name{test_base}
\alias{test_base}
\title{test base}
\arguments{
\item{tail}{specify either \code{1} or \code{2} tail}

\item{conf}{the confidence}
}
\value{
the list of stats.
}
\description{
Calculates the statistics, prints in a nice format,
and returns a list for easy access of the stats.
}
