% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{two_prop_assumptions}
\alias{two_prop_assumptions}
\title{Two Proportions t-test Assumptions}
\usage{
two_prop_assumptions()
}
\description{
Prints out the assumptions.
}
\examples{
two_prop_assumptions()
}
