% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample.R
\name{two_prop_se}
\alias{two_prop_se}
\title{Two Proportions Standard Error}
\usage{
two_prop_se(phat1, n1, phat2, n2)
}
\arguments{
\item{phat1}{1st sample proportion}

\item{n1}{1st sample size}

\item{phat2}{2nd sample proportion}

\item{n2}{2nd sample size}
}
\description{
Two Proportions Standard Error
}
