% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample.R
\name{two_prop_z}
\alias{two_prop_z}
\title{Two Proportions z-test}
\usage{
two_prop_z(phat1, n1, phat2, n2, tail, conf = 0.95)
}
\arguments{
\item{phat1}{1st sample proportion}

\item{n1}{1st sample size}

\item{phat2}{2nd sample proportion}

\item{n2}{2nd sample size}

\item{tail}{specify either \code{1} or \code{2} tail}

\item{conf}{the confidence}
}
\value{
the list of stats.
}
\description{
Calculates the statistics, prints in a nice format,
and returns a list for easy access of the stats.
}
\seealso{
\code{\link[stats:prop.test]{stats::prop.test()}}

\code{prop.test(table(data$X, data$Y))}
}
