% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample.R
\name{two_sample_se}
\alias{two_sample_se}
\title{Two Sample Standard Error}
\usage{
two_sample_se(s1, n1, s2, n2)
}
\arguments{
\item{s1}{1st sample SD}

\item{n1}{1st sample size}

\item{s2}{2nd sample SD}

\item{n2}{2nd sample size}
}
\value{
the two sample se(x1 - x2)
}
\description{
Two Sample Standard Error
}
