% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-square.R
\name{xsq_test}
\alias{xsq_test}
\title{Chi-Square X^2-test}
\usage{
xsq_test(xsq, df)
}
\arguments{
\item{xsq}{the x^2 stat}

\item{df}{the degrees of freedom}
}
\value{
the p value
}
\description{
Chi-Square X^2-test
}
