% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{z_crit}
\alias{z_crit}
\title{Z Critical}
\usage{
z_crit(conf)
}
\arguments{
\item{conf}{The confidence as a decimal.}
}
\value{
the Z Critical value
}
\description{
Calculates the Z Critical value from a confidence.
}
\examples{
z_crit(0.95) # 95\% Confident
z_crit(0.90) # 90\% Confident
}
