% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{z_test}
\alias{z_test}
\title{z-test}
\usage{
z_test(z, tail)
}
\arguments{
\item{z}{the z stat}

\item{tail}{specify either \code{1} or \code{2} tail}
}
\value{
p value
}
\description{
Performs a z-test.
}
\seealso{
\code{\link[stats:Normal]{stats::pnorm()}}
}
