% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{generate_frame}
\alias{generate_frame}
\title{Generate data.frame}
\usage{
generate_frame(..., as_factors = TRUE)
}
\arguments{
\item{...}{The observed frequencies (fo) in the form of \cr
Col1 = c(Row1 = 52, Row2 = 61), Col2 = c(Row1 = 22, Row2 = 39)}

\item{as_factors}{Boolean specifying whether to treat the Cols and Rows as factors}
}
\value{
the new data.frame
}
\description{
Generates a data.frame from observed frequencies.
}
\examples{
frame1 <- generate_frame(
  Nicotine = c(No = 148, Yes = 52),
  Placebo = c(No = 182, Yes = 18)
)
addmargins(table(frame1$Rows, frame1$Cols))

frame2 <- generate_frame(
  Home = c("2018" = 391, "2020" = 454),
  Shared = c("2018" = 185, "2020" = 139)
)
addmargins(table(frame2$Rows, frame2$Cols))
}
