% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{logistic_moe}
\alias{logistic_moe}
\title{Logistic Regression Margin of Error}
\usage{
logistic_moe(se, conf = 0.95)
}
\arguments{
\item{se}{The standard error}

\item{conf}{The confidence}
}
\value{
The margin of Error
}
\description{
Logistic Regression is a z-distribution.
}
