% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-regression.R
\name{lr_assumptions}
\alias{lr_assumptions}
\title{Linear Regression Assumptions}
\usage{
lr_assumptions()
}
\description{
Prints out the assumptions.
}
\seealso{
\code{stats::lm(Y ~ X, data)}

\code{stats::lm(Y ~ X1 * X2, data) # Interaction}

\code{stats::lm(Y ~ X1 + X2, data) # No Interaction}
}
